Implement _DelayedCall for tail call optimization.
Applying a Lisp function is now equivalent to creating a _DelayedCall
object that represents the call to be made.  The _DelayedCall is passed up the Python call stack until it can it reaches the _eval function of a forcing (non-tail) context.  At that point, only once the call stack has been shortened as much as possible, the _DelayedCall is forced.

Forcing the _DelayedCall might result in another _DelayedCall, if the forced function itself had a tail call.  In that case, the forcing context would need to force the new _DelayedCall.  This process would be repeated until a value were finally obtained.  Or, it might repeat indefinitely.  The important fact is that this repetition is done iteratively, so the Python call stack doesn't fill up.