Let the formatter print Python lists as Lisp lists.
This may be unnecessary in the long term.  But, we have not yet
decided whether Python lists should ever be passed around.  It's
possible that we'll only want to ever pass Lisp lists through the
interpreter.  If that's the case, then we can revert this change as
soon as we enforce that the interpreter never touches Python lists.